Implement winning the game
This commit implements a method which is executed on each turn to determine whether a player has won. If a player has won, that player is displayed as the winner and the game ends.

Also, if a square has already been clicked on, the turn will not progress if it is clicked on again.